Remove cursor in SP
