+++ packages/service-provider-browser/src/unsupported-cursor.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { Document, Cursor } from '@mongosh/service-provider-core';
+import { Document, Cursor, CursorFlag } from '@mongosh/service-provider-core';
+import { MongoshUnimplementedError } from '@mongosh/errors';
 
 /**
  * Defines a cursor for an unsupported operation.
@@ -16,124 +17,136 @@ class UnsupportedCursor implements Cursor {
     this.message = message;
   }
 
+  addCursorFlag(flag: CursorFlag, value: boolean): any {
+    throw new MongoshUnimplementedError('Method not implemented');
+  }
+
+  noServiceProviderCursorTimeout(): any {
+    throw new MongoshUnimplementedError('Method not implemented');
+  }
+
+  setReadPreference(mode: any): Cursor {
+    throw new MongoshUnimplementedError('Method not implemented');
+  }
+
   addOption(option: number): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   allowPartialResults(): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   batchSize(size: number): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   close(options: Document): Promise<void> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   clone(): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   isClosed(): boolean {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   collation(spec: Document): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   comment(cmt: string): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   count(): Promise<number> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   forEach(f: any): Promise<void> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   hasNext(): Promise<boolean> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   hint(index: string): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   isExhausted(): Promise<boolean> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   itcount(): Promise<number> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   limit(value: number): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   map(f: any): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   max(indexBounds: Document): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   maxTimeMS(value: number): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   min(indexBounds: Document): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   next(): Promise<any> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   noCursorTimeout(): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   oplogReplay(): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
-  projection(spec: Document): Cursor {
-    throw new Error('Method not implemented.');
+  project(spec: Document): Cursor {
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   readPref(preference: string, tagSet?: Document[]): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   returnKey(enabled: boolean): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   size(): Promise<number> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   skip(value: number): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   sort(spec: Document): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   tailable(): Cursor {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   explain(verbosity: string): Promise<any> {
-    throw new Error('Method not implemented.');
+    throw new MongoshUnimplementedError('Method not implemented.');
   }
 
   /**
@@ -143,7 +156,7 @@ class UnsupportedCursor implements Cursor {
    * @returns {Promise} The rejected promise.
    */
   async toArray(): Promise<Document[]> {
-    throw new Error(this.message);
+    throw new MongoshUnimplementedError(this.message);
   }
 }
